add endpoint for delete on resource category
