+++ docs/resourcesCategory/resourcesCategory.yaml
@@ -211,4 +211,61 @@ paths:
               example:
                 status: 404
                 code: NOT_FOUND
-                message: Resource Category with the specified id was not found.
+                message: Resource Category with the specified id was not found.    delete:
+      security:
+        - bearerAuth: []
+      tags:
+        - ResourcesCategories
+      summary: Delete ResourcesCategory by ID
+      description: Finds and deletes an ResourcesCategory with the specified ID.
+      produces:
+        - application/json
+      parameters:
+        - name: id
+          in: path
+          required: true
+          description: ID of the ResourcesCategory that needs to be deleted
+          type: string
+      responses:
+        204:
+          description: No Content
+        400:
+          description: Bad Request
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/Error'
+              example:
+                status: 400
+                code: INVALID_ID
+                message: ID is invalid.
+        401:
+          description: Unauthorized
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/Error'
+              example:
+                status: 401
+                code: UNAUTHORIZED
+                message: The requested URL requires user authorization.
+        403:
+          description: Forbidden
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/Error'
+              example:
+                status: 403
+                code: FORBIDDEN
+                message: You do not have permission to perform this action.
+        404:
+          description: Not Found
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/Error'
+              example:
+                status: 404
+                code: DOCUMENT_NOT_FOUND
+                message: ResourcesCategory with the specified id was not found.
