+++ routes/resourcesCategory.js
@@ -3,12 +3,12 @@ const router = require('express').Router()
 const ResourceCategory = require('~/models/resourcesCategory')
 const asyncWrapper = require('~/middlewares/asyncWrapper')
 const { authMiddleware, restrictTo } = require('~/middlewares/auth')
+const isEntityValid = require('~/middlewares/entityValidation')
 const resourcesCategoryController = require('~/controllers/resourcesCategory')
 
 const {
   roles: { TUTOR }
 } = require('~/consts/auth')
-const isEntityValid = require('~/middlewares/entityValidation')
 
 const params = [{ model: ResourceCategory, idName: 'id' }]
 
@@ -18,5 +18,6 @@ router.get('/', asyncWrapper(resourcesCategoryController.getResourcesCategories)
 router.get('/names', asyncWrapper(resourcesCategoryController.getResourcesCategoriesNames))
 router.post('/', asyncWrapper(resourcesCategoryController.createResourcesCategory))
 router.patch('/:id', isEntityValid({ params }), asyncWrapper(resourcesCategoryController.updateResourceCategory))
+router.delete('/:id', isEntityValid({ params }), asyncWrapper(resourcesCategoryController.deleteResourceCategory))
 
 module.exports = router
